SiteUser 엔티티 추가, 다른 엔티티의 주키를 Long 타입으로

+++ src/main/java/com/ll/exam/sbb/answer/AnswerController.java
@@ -22,7 +22,7 @@ public class AnswerController {
 
     @PostMapping("/create/{id}")
     public String createAnswer(
-            Model model, @PathVariable("id") Integer id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
+            Model model, @PathVariable("id") long id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
         Question question = this.questionService.getQuestion(id);
 
         if (bindingResult.hasErrors()) {
